# Add perf test for logging v2