--- a/v2/tests/CMakeLists.txt
+++ b/v2/tests/CMakeLists.txt
@@ -29,4 +29,5 @@ if(${run_int_tests})
 endif()
 
 if(${run_perf_tests})
+   add_subdirectory(logger_perf)
 endif()