--- a/v2/tests/CMakeLists.txt
+++ b/v2/tests/CMakeLists.txt
@@ -29,5 +29,7 @@ if(${run_int_tests})
 endif()
 
 if(${run_perf_tests})
-   add_subdirectory(logger_perf)
+   if(WIN32)
+       add_subdirectory(logger_perf)
+   endif()
 endif()