--- a/v2/tests/logger_perf/timer_win32.c
+++ b/v2/tests/logger_perf/timer_win32.c
@@ -1,4 +1,5 @@
 // Copyright (C) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 #include <stdio.h>