… v2/tests/logger_perf/CMakeLists.txt