Test summary:
=============
results[LOGGER_TYPE_V1 (0)][CONSUMER_DISABLED (0)]: 4350000 logs in 5 s (5006098 us), 868940 logs/s
results[LOGGER_TYPE_V1 (0)][CONSUMER_ENABLED (1)]: 3720000 logs in 5 s (5006130 us), 743089 logs/s
results[LOGGER_TYPE_V2_NO_CONTEXT (1)][CONSUMER_DISABLED (0)]: 4350000 logs in 5 s (5010258 us), 868219 logs/s
results[LOGGER_TYPE_V2_NO_CONTEXT (1)][CONSUMER_ENABLED (1)]: 3730000 logs in 5 s (5002687 us), 745599 logs/s
results[LOGGER_TYPE_V2_LOCAL_CONTEXT_CREATED_ONCE (2)][CONSUMER_DISABLED (0)]: 3940000 logs in 5 s (5008685 us), 786634 logs/s
results[LOGGER_TYPE_V2_LOCAL_CONTEXT_CREATED_ONCE (2)][CONSUMER_ENABLED (1)]: 3400000 logs in 5 s (5014230 us), 678070 logs/s
results[LOGGER_TYPE_V2_DYNAMIC_CONTEXT_CREATED_ONCE (3)][CONSUMER_DISABLED (0)]: 3830000 logs in 5 s (5004486 us), 765313 logs/s
results[LOGGER_TYPE_V2_DYNAMIC_CONTEXT_CREATED_ONCE (3)][CONSUMER_ENABLED (1)]: 3320000 logs in 5 s (5014516 us), 662078 logs/s
results[LOGGER_TYPE_V2_LOCAL_CONTEXT_CREATED_EVERY_TIME (4)][CONSUMER_DISABLED (0)]: 3520000 logs in 5 s (5004041 us), 703431 logs/s
results[LOGGER_TYPE_V2_LOCAL_CONTEXT_CREATED_EVERY_TIME (4)][CONSUMER_ENABLED (1)]: 3140000 logs in 5 s (5013909 us), 626258 logs/s
results[LOGGER_TYPE_V2_DYNAMIC_CONTEXT_CREATED_EVERY_TIME (5)][CONSUMER_DISABLED (0)]: 3200000 logs in 5 s (5000089 us), 639989 logs/s
results[LOGGER_TYPE_V2_DYNAMIC_CONTEXT_CREATED_EVERY_TIME (5)][CONSUMER_ENABLED (1)]: 2940000 logs in 5 s (5014051 us), 586352 logs/s